Get books locally instead of api call
